+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 
 ### Documentation
 
+- 2026-09-23 | 📝 docs(design): define accessible design system
+
 - 2026-09-22 | 📝 docs: consolidate physical purchase validation
 
 - 2026-09-22 | 📝 docs: consolidate hosted validation evidence